app: plugin-management: Fix downloading extra files on windows
So that there is .exe on the end of them on windows.

+++ app/electron/plugin-management.ts
@@ -633,8 +633,18 @@ async function downloadExtraFiles(
       if (!value.output || !value.input || value.input === value.output) {
         continue;
       }
-      const outputFile = path.join(binDir, value.output);
+      let outputFile = path.join(binDir, value.output);
+      // If on Windows, ensure that the output file ends with .exe
+      // For example, minikube should be minikube.exe
+      // If the extra file is a .js file, we do not add .exe
+      if (os.platform() === 'win32' && !value.output.endsWith('.js')) {
+        outputFile = path.join(binDir, value.output) + '.exe';
+      }
+
       const inputFile = path.join(binDir, value.input);
+      if (inputFile === outputFile) {
+        continue;
+      }
 
       fs.copyFileSync(inputFile, outputFile);
       fs.rmSync(inputFile);
